backend/requirements.txt
aiohttp>=3.12.15
fastapi>=0.115.8
uvicorn>=0.34.0
python-dotenv
asyncpg
pydantic
